fix: solved some sonarcloud issues and changed false positives calculation

+++ src/control/include/node_/node_control.hpp
@@ -19,15 +19,15 @@
 struct ControlParameters {
   bool using_simulated_se_;
   bool use_simulated_planning_;
-  double lookahead_gain_;
-  double pid_kp_;
-  double pid_ki_;
-  double pid_kd_;
-  double pid_tau_;
-  double pid_t_;
-  double pid_lim_min_;
-  double pid_lim_max_;
-  double pid_anti_windup_;
+  long double lookahead_gain_;
+  long double pid_kp_;
+  long double pid_ki_;
+  long double pid_kd_;
+  long double pid_tau_;
+  long double pid_t_;
+  long double pid_lim_min_;
+  long double pid_lim_max_;
+  long double pid_anti_windup_;
 };
 
 /**
